Deleted the old project and copied over the tests to the new project.
