--- EnumStringValueTests/packages.config
@@ -1,5 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="FluentAssertions" version="3.3.0" targetFramework="net451" />
-  <package id="NUnit" version="2.6.3" targetFramework="net45" />
-</packages>
