+++ EnumStringValues.Tests/packages.config
@@ -1,5 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
+  <package id="FluentAssertions" version="4.19.4" targetFramework="net461" />
   <package id="MSTest.TestAdapter" version="1.1.18" targetFramework="net461" />
   <package id="MSTest.TestFramework" version="1.1.18" targetFramework="net461" />
+  <package id="NUnit" version="3.9.0" targetFramework="net461" />
 </packages>
