+++ EnumStringValues.sln
@@ -3,8 +3,6 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.27004.2010
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EnumStringValueTests", "EnumStringValueTests\EnumStringValueTests.csproj", "{6A5C3CB5-9925-4D8C-84A5-F368FF4C5CFB}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "EnumStringValues", "EnumStringValues\EnumStringValues.csproj", "{EB9B3169-1860-4A12-B342-677ACCFFB64F}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EnumStringValues.Tests", "EnumStringValues.Tests\EnumStringValues.Tests.csproj", "{31EA3C2F-F0CD-4A73-88F6-244952B725E9}"
@@ -15,10 +13,6 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{6A5C3CB5-9925-4D8C-84A5-F368FF4C5CFB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{6A5C3CB5-9925-4D8C-84A5-F368FF4C5CFB}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{6A5C3CB5-9925-4D8C-84A5-F368FF4C5CFB}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{6A5C3CB5-9925-4D8C-84A5-F368FF4C5CFB}.Release|Any CPU.Build.0 = Release|Any CPU
 		{EB9B3169-1860-4A12-B342-677ACCFFB64F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{EB9B3169-1860-4A12-B342-677ACCFFB64F}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{EB9B3169-1860-4A12-B342-677ACCFFB64F}.Release|Any CPU.ActiveCfg = Release|Any CPU
